Refactor package name to spark3 so that both plugins can co-exist
