+++ pinot-plugins/pinot-batch-ingestion/pinot-batch-ingestion-spark-3.2/src/main/java/org/apache/pinot/plugin/ingestion/batch/spark3/SparkSegmentMetadataPushJobRunner.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.plugin.ingestion.batch.spark;
+package org.apache.pinot.plugin.ingestion.batch.spark3;
 
 import java.io.File;
 import java.io.IOException;
@@ -24,8 +24,8 @@ import java.io.Serializable;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
 import org.apache.pinot.segment.local.utils.SegmentPushUtils;
 import org.apache.pinot.spi.env.PinotConfiguration;
 import org.apache.pinot.spi.filesystem.PinotFS;
@@ -43,22 +43,19 @@ import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.api.java.function.VoidFunction;
 
 
-public class SparkSegmentUriPushJobRunner implements IngestionJobRunner, Serializable {
+public class SparkSegmentMetadataPushJobRunner implements IngestionJobRunner, Serializable {
   private SegmentGenerationJobSpec _spec;
 
-  public SparkSegmentUriPushJobRunner() {
+  public SparkSegmentMetadataPushJobRunner() {
   }
 
-  public SparkSegmentUriPushJobRunner(SegmentGenerationJobSpec spec) {
+  public SparkSegmentMetadataPushJobRunner(SegmentGenerationJobSpec spec) {
     init(spec);
   }
 
   @Override
   public void init(SegmentGenerationJobSpec spec) {
     _spec = spec;
-    if (_spec.getPushJobSpec() == null) {
-      throw new RuntimeException("Missing PushJobSpec");
-    }
   }
 
   @Override
@@ -69,7 +66,7 @@ public class SparkSegmentUriPushJobRunner implements IngestionJobRunner, Seriali
       PinotFSFactory.register(pinotFSSpec.getScheme(), pinotFSSpec.getClassName(), new PinotConfiguration(pinotFSSpec));
     }
 
-    //Get outputFS for writing output Pinot segments
+    //Get outputFS for writing output pinot segments
     URI outputDirURI;
     try {
       outputDirURI = new URI(_spec.getOutputDirURI());
@@ -80,7 +77,6 @@ public class SparkSegmentUriPushJobRunner implements IngestionJobRunner, Seriali
       throw new RuntimeException("outputDirURI is not valid - '" + _spec.getOutputDirURI() + "'");
     }
     PinotFS outputDirFS = PinotFSFactory.create(outputDirURI.getScheme());
-
     //Get list of files to process
     String[] files;
     try {
@@ -88,44 +84,45 @@ public class SparkSegmentUriPushJobRunner implements IngestionJobRunner, Seriali
     } catch (IOException e) {
       throw new RuntimeException("Unable to list all files under outputDirURI - '" + outputDirURI + "'");
     }
-    List<String> segmentUris = new ArrayList<>();
+
+    List<String> segmentsToPush = new ArrayList<>();
     for (String file : files) {
-      URI uri = URI.create(file);
-      if (uri.getPath().endsWith(Constants.TAR_GZ_FILE_EXT)) {
-        URI updatedURI = SegmentPushUtils
-            .generateSegmentTarURI(outputDirURI, uri, _spec.getPushJobSpec().getSegmentUriPrefix(),
-                _spec.getPushJobSpec().getSegmentUriSuffix());
-        segmentUris.add(updatedURI.toString());
+      if (file.endsWith(Constants.TAR_GZ_FILE_EXT)) {
+        segmentsToPush.add(file);
       }
     }
 
     int pushParallelism = _spec.getPushJobSpec().getPushParallelism();
     if (pushParallelism < 1) {
-      pushParallelism = segmentUris.size();
+      pushParallelism = segmentsToPush.size();
     }
     if (pushParallelism == 1) {
       // Push from driver
       try {
-        SegmentPushUtils.sendSegmentUris(_spec, segmentUris);
+        SegmentPushUtils.pushSegments(_spec, outputDirFS, segmentsToPush);
       } catch (RetriableOperationException | AttemptsExceededException e) {
         throw new RuntimeException(e);
       }
     } else {
       JavaSparkContext sparkContext = JavaSparkContext.fromSparkContext(SparkContext.getOrCreate());
-      JavaRDD<String> pathRDD = sparkContext.parallelize(segmentUris, pushParallelism);
+      JavaRDD<String> pathRDD = sparkContext.parallelize(segmentsToPush, pushParallelism);
+      URI finalOutputDirURI = outputDirURI;
       // Prevent using lambda expression in Spark to avoid potential serialization exceptions, use inner function
       // instead.
       pathRDD.foreach(new VoidFunction<String>() {
         @Override
-        public void call(String segmentUri)
+        public void call(String segmentTarPath)
             throws Exception {
+          PluginManager.get().init();
+          for (PinotFSSpec pinotFSSpec : pinotFSSpecs) {
+            PinotFSFactory
+                .register(pinotFSSpec.getScheme(), pinotFSSpec.getClassName(), new PinotConfiguration(pinotFSSpec));
+          }
           try {
-            PluginManager.get().init();
-            for (PinotFSSpec pinotFSSpec : pinotFSSpecs) {
-              PinotFSFactory
-                  .register(pinotFSSpec.getScheme(), pinotFSSpec.getClassName(), new PinotConfiguration(pinotFSSpec));
-            }
-            SegmentPushUtils.sendSegmentUris(_spec, Arrays.asList(segmentUri));
+            Map<String, String> segmentUriToTarPathMap = SegmentPushUtils
+                .getSegmentUriToTarPathMap(finalOutputDirURI, _spec.getPushJobSpec(), new String[]{segmentTarPath});
+            SegmentPushUtils.sendSegmentUriAndMetadata(_spec, PinotFSFactory.create(finalOutputDirURI.getScheme()),
+                segmentUriToTarPathMap);
           } catch (RetriableOperationException | AttemptsExceededException e) {
             throw new RuntimeException(e);
           }
