+++ pinot-plugins/pinot-batch-ingestion/pinot-batch-ingestion-spark-3.2/src/main/java/org/apache/pinot/plugin/ingestion/batch/spark3/SparkSegmentUriPushJobRunner.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.plugin.ingestion.batch.spark;
+package org.apache.pinot.plugin.ingestion.batch.spark3;
 
 import java.io.File;
 import java.io.IOException;
@@ -43,19 +43,22 @@ import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.api.java.function.VoidFunction;
 
 
-public class SparkSegmentTarPushJobRunner implements IngestionJobRunner, Serializable {
+public class SparkSegmentUriPushJobRunner implements IngestionJobRunner, Serializable {
   private SegmentGenerationJobSpec _spec;
 
-  public SparkSegmentTarPushJobRunner() {
+  public SparkSegmentUriPushJobRunner() {
   }
 
-  public SparkSegmentTarPushJobRunner(SegmentGenerationJobSpec spec) {
+  public SparkSegmentUriPushJobRunner(SegmentGenerationJobSpec spec) {
     init(spec);
   }
 
   @Override
   public void init(SegmentGenerationJobSpec spec) {
     _spec = spec;
+    if (_spec.getPushJobSpec() == null) {
+      throw new RuntimeException("Missing PushJobSpec");
+    }
   }
 
   @Override
@@ -66,7 +69,7 @@ public class SparkSegmentTarPushJobRunner implements IngestionJobRunner, Seriali
       PinotFSFactory.register(pinotFSSpec.getScheme(), pinotFSSpec.getClassName(), new PinotConfiguration(pinotFSSpec));
     }
 
-    //Get outputFS for writing output pinot segments
+    //Get outputFS for writing output Pinot segments
     URI outputDirURI;
     try {
       outputDirURI = new URI(_spec.getOutputDirURI());
@@ -77,6 +80,7 @@ public class SparkSegmentTarPushJobRunner implements IngestionJobRunner, Seriali
       throw new RuntimeException("outputDirURI is not valid - '" + _spec.getOutputDirURI() + "'");
     }
     PinotFS outputDirFS = PinotFSFactory.create(outputDirURI.getScheme());
+
     //Get list of files to process
     String[] files;
     try {
@@ -84,43 +88,44 @@ public class SparkSegmentTarPushJobRunner implements IngestionJobRunner, Seriali
     } catch (IOException e) {
       throw new RuntimeException("Unable to list all files under outputDirURI - '" + outputDirURI + "'");
     }
-
-    List<String> segmentsToPush = new ArrayList<>();
+    List<String> segmentUris = new ArrayList<>();
     for (String file : files) {
-      if (file.endsWith(Constants.TAR_GZ_FILE_EXT)) {
-        segmentsToPush.add(file);
+      URI uri = URI.create(file);
+      if (uri.getPath().endsWith(Constants.TAR_GZ_FILE_EXT)) {
+        URI updatedURI = SegmentPushUtils
+            .generateSegmentTarURI(outputDirURI, uri, _spec.getPushJobSpec().getSegmentUriPrefix(),
+                _spec.getPushJobSpec().getSegmentUriSuffix());
+        segmentUris.add(updatedURI.toString());
       }
     }
 
     int pushParallelism = _spec.getPushJobSpec().getPushParallelism();
     if (pushParallelism < 1) {
-      pushParallelism = segmentsToPush.size();
+      pushParallelism = segmentUris.size();
     }
     if (pushParallelism == 1) {
       // Push from driver
       try {
-        SegmentPushUtils.pushSegments(_spec, outputDirFS, segmentsToPush);
+        SegmentPushUtils.sendSegmentUris(_spec, segmentUris);
       } catch (RetriableOperationException | AttemptsExceededException e) {
         throw new RuntimeException(e);
       }
     } else {
       JavaSparkContext sparkContext = JavaSparkContext.fromSparkContext(SparkContext.getOrCreate());
-      JavaRDD<String> pathRDD = sparkContext.parallelize(segmentsToPush, pushParallelism);
-      URI finalOutputDirURI = outputDirURI;
+      JavaRDD<String> pathRDD = sparkContext.parallelize(segmentUris, pushParallelism);
       // Prevent using lambda expression in Spark to avoid potential serialization exceptions, use inner function
       // instead.
       pathRDD.foreach(new VoidFunction<String>() {
         @Override
-        public void call(String segmentTarPath)
+        public void call(String segmentUri)
             throws Exception {
-          PluginManager.get().init();
-          for (PinotFSSpec pinotFSSpec : pinotFSSpecs) {
-            PinotFSFactory
-                .register(pinotFSSpec.getScheme(), pinotFSSpec.getClassName(), new PinotConfiguration(pinotFSSpec));
-          }
           try {
-            SegmentPushUtils.pushSegments(_spec, PinotFSFactory.create(finalOutputDirURI.getScheme()),
-                Arrays.asList(segmentTarPath));
+            PluginManager.get().init();
+            for (PinotFSSpec pinotFSSpec : pinotFSSpecs) {
+              PinotFSFactory
+                  .register(pinotFSSpec.getScheme(), pinotFSSpec.getClassName(), new PinotConfiguration(pinotFSSpec));
+            }
+            SegmentPushUtils.sendSegmentUris(_spec, Arrays.asList(segmentUri));
           } catch (RetriableOperationException | AttemptsExceededException e) {
             throw new RuntimeException(e);
           }
