+++ pinot-plugins/pinot-batch-ingestion/pinot-batch-ingestion-spark-3.2/src/main/resources/segmentTarPushJobSpec.yaml
@@ -19,10 +19,10 @@
 
 executionFrameworkSpec:
   name: 'spark'
-  segmentGenerationJobRunnerClassName: 'org.apache.pinot.plugin.ingestion.batch.spark.SparkSegmentGenerationJobRunner'
-  segmentTarPushJobRunnerClassName: 'org.apache.pinot.plugin.ingestion.batch.spark.SparkSegmentTarPushJobRunner'
-  segmentUriPushJobRunnerClassName: 'org.apache.pinot.plugin.ingestion.batch.spark.SparkSegmentUriPushJobRunner'
-  segmentMetadataPushJobRunnerClassName: 'org.apache.pinot.plugin.ingestion.batch.spark.SparkSegmentMetadataPushJobRunner'
+  segmentGenerationJobRunnerClassName: 'SparkSegmentGenerationJobRunner'
+  segmentTarPushJobRunnerClassName: 'SparkSegmentTarPushJobRunner'
+  segmentUriPushJobRunnerClassName: 'SparkSegmentUriPushJobRunner'
+  segmentMetadataPushJobRunnerClassName: 'SparkSegmentMetadataPushJobRunner'
 jobType: SegmentTarPush
 inputDirURI: 'file:///path/to/input'
 includeFileNamePattern: 'glob:**/*.parquet'
